Move delete purchase into dropdown
