+++ components/Purchases/data-table.tsx
@@ -18,19 +18,24 @@ import {
     TableHeader,
     TableRow,
 } from "@/components/ui/table";
+import { Purchase } from "./columns";
 
 interface DataTableProps<TData, TValue> {
     columns: ColumnDef<TData, TValue>[]
     data: TData[],
     editComponentOpen: boolean,
-    handleEditComponent: (open: boolean) => void
+    handleEditComponent: (open: boolean) => void,
+    deletePurchase: (id: bigint) => void,
+    setCurrentPurchase: (purchase: Purchase) => void
 }
 
 export function DataTable<TData, TValue>({
     columns,
     data,
     editComponentOpen,
-    handleEditComponent
+    handleEditComponent,
+    deletePurchase,
+    setCurrentPurchase
   }: DataTableProps<TData, TValue>) {
     const [sorting, setSorting] = React.useState<SortingState>([])
 
@@ -45,7 +50,9 @@ export function DataTable<TData, TValue>({
     },
       meta: {
         editComponentOpen,
-        handleEditComponent
+        handleEditComponent,
+        deletePurchase,
+        setCurrentPurchase
       }
     })
