+++ components/Purchases/index.tsx
@@ -7,8 +7,13 @@ import dayjs from 'dayjs';
 import isoWeek from 'dayjs/plugin/isoWeek'
 import * as _ from "lodash";
 import { useContext, useState } from "react";
+import { useRouter } from "next/navigation";
+import EditPurchase from "../EditPurchase";
 
 dayjs.extend(isoWeek);
+
+// @ts-ignore
+BigInt.prototype.toJSON = function() { return this.toString() }
 
 interface PurchaseProps {
     purchases: Array<Purchase>
@@ -38,7 +43,30 @@ const filterPurchases = (purchases: Array<Purchase>, filter: string) => {
 }
 
 export default function Purchases(props: PurchaseProps) {
+
+    const router = useRouter();
+
+    async function deletePurchase(id: bigint) {
+        const data = {
+            id
+        }
+    
+        try {
+            await fetch('/api/purchase', { 
+                method: 'DELETE',
+                headers: {
+                    'Content-Type': 'application/json'
+                },
+                body: JSON.stringify(data)
+            });
+            router.refresh();
+        } catch (error) {
+            console.error(error);
+        }
+    }
+
     const [editComponentOpen, setEditComponentOpen] = useState(false);
+    const [currentPurchase, setCurrentPurchase] = useState({} as Purchase);
 
     const { purchases } = props;
 
@@ -51,7 +79,15 @@ export default function Purchases(props: PurchaseProps) {
     return (
         <>
             <div className="container mx-auto text-center mb-10 mt-10">
-                <DataTable columns={columns} data={sortedPurchases} editComponentOpen={editComponentOpen} handleEditComponent={setEditComponentOpen}  />
+                <DataTable 
+                    columns={columns} 
+                    data={sortedPurchases} 
+                    editComponentOpen={editComponentOpen} 
+                    handleEditComponent={setEditComponentOpen}
+                    deletePurchase={deletePurchase}
+                    setCurrentPurchase={setCurrentPurchase}
+                />
+                <EditPurchase purchase={currentPurchase} open={editComponentOpen} openOnChange={setEditComponentOpen} />
             </div>
         </>
     )
